clang-tidy config + CI lint gate (M197)
Adds a .clang-tidy configuration and wires two new CI jobs, plus fixes the
genuine issues clang-tidy surfaced in the engine's own sources.

- .clang-tidy: a focused, high-signal check set (bugprone / performance /
  portability / misc + a curated readability + modernize slice). Opinionated or
  house-style-clashing checks are disabled explicitly (const-correctness,
  enum-size, move-const-arg — the last gives compile-breaking advice for the
  engine's rvalue-ref texture sinks). Scoped to first-party headers via
  HeaderFilterRegex; strict in CI (WarningsAsErrors '*').
- CI: a `tidy` job runs clang-tidy over every engine/src/**/*.cpp on push/PR and
  fails on any diagnostic; a `sanitizers` job builds the unit suite with
  -DMAZ_SANITIZE=address,undefined and runs it under ASan+UBSan+LeakSanitizer
  (leveraging M196).
- Source fixes for the real finds so the gate is genuinely green (verified: zero
  diagnostics across all engine sources): removed a dead `using LogLevel`,
  merged a cloned switch branch (SDL quit/close), dropped a misplaced const on a
  VkPipeline handle, made an int->size_t multiply explicit in the image loader,
  and collapsed a redundant boolean return in SsaoPass.

Verified: engine + full app build clean (warnings-as-errors), unit suite green,
and clang-tidy reports no diagnostics over the first-party sources.

+++ engine/src/render/Model.cpp
@@ -136,7 +136,7 @@ bool decodeImage(const cgltf_image* image, const char* gltfPath, std::vector<uin
     }
     outW = static_cast<uint32_t>(w);
     outH = static_cast<uint32_t>(h);
-    outPx.assign(pixels, pixels + static_cast<size_t>(w) * h * 4);
+    outPx.assign(pixels, pixels + static_cast<size_t>(w) * static_cast<size_t>(h) * 4);
     stbi_image_free(pixels);
     return true;
 }
